[setting] #3 - spotless Gradle 플러그인 추가 & Java 포맷 규칙 추가, Git pre-commit 훅 추가 (spotlessApply + compileJava 자동 실행)
