+++ src/main/java/com/example/coalawebbackend/common/config/SwaggerConfig.java
@@ -1,0 +1,30 @@
+package com.example.coalawebbackend.common.config;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.servers.Server;
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@OpenAPIDefinition(
+        servers = {
+                @Server(url = "http://localhost:8080", description = "로컬 서버")
+        }
+)
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public OpenAPI openAPI() {
+        Info info = new Info()
+                .title("Coala Web Backend API")
+                .version("v1.0.0")
+                .description("Coala Community Project");
+
+        return new OpenAPI()
+                .components(new Components())
+                .info(info);
+    }
+}
