--- src/main/java/com/example/coalawebbackend/global/config/SwaggerConfig.java
@@ -1,20 +1,0 @@
-package com.example.coalawebbackend.global.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("Coala Web Backend API")
-                        .description("Coala Web Backend API 명세서")
-                        .version("v1.0.0"));
-    }
-}
-
